First pass at using redis scripting for presence
Long story short: We store the users attached to
a particular process in a list. We use a redis
script to reconcile the actual user count on an
interval or when a user joins/leaves or when
a server starts up/shuts down/crashes. The process
list is also on a TTL, so it should never be more
than 15 seconds out of date.